+++ stories/static/data/languages.js
@@ -1,0 +1,1 @@
+export const LANGUAGES = ['da', 'de', 'fr', 'en', 'es', 'fi', 'it', 'nl', 'pt', 'pl', 'sv'];