+++ src/components/menu/Menu.js
@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React, { createRef, PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import cx from 'classnames';
@@ -19,9 +19,9 @@ const POSITION = {
 
 class Menu extends PureComponent {
   state = {};
-
+  menuNode = createRef();
   componentDidMount() {
-    const { width, height } = this.menuNode.getBoundingClientRect();
+    const { width, height } = this.menuNode.current.getBoundingClientRect();
     this.setState({ width, height });
 
     const { position } = this.props;
@@ -61,7 +61,7 @@ class Menu extends PureComponent {
       this.setState({ position: this.calculatePosition() });
     }
 
-    const { width, height } = this.menuNode.getBoundingClientRect();
+    const { width, height } = this.menuNode.current.getBoundingClientRect();
     if (prevState.width !== width || prevState.height !== height) {
       this.setState({ width, height }); // eslint-disable-line
     }
@@ -229,16 +229,8 @@ class Menu extends PureComponent {
 
     return (
       <div data-teamleader-ui="menu" className={classNames} style={this.getRootStyle()} {...others}>
-        {outline ? (
-          <div className={cx(uiUtilities['box-shadow-200'], theme['outline'])} style={{ width, height }} />
-        ) : null}
-        <ul
-          ref={node => {
-            this.menuNode = node;
-          }}
-          className={theme['menu-inner']}
-          style={this.getMenuStyle()}
-        >
+        {outline ? <div className={theme['outline']} style={{ width, height }} /> : null}
+        <ul ref={this.menuNode} className={theme['menu-inner']} style={this.getMenuStyle()}>
           {this.getItems()}
         </ul>
       </div>